Training on Real Objects
Created two seperate ways to train this
The first was trained specifically for the set: to be ableto indentify
one of 29 images
(29 objects not 30 in the 30objects.zip ha)

The second way takes in two images and outputs either match or no-match.
Training the network this way allows new 'unseen' objects to be added
and the network can adapt easily.